pyproject.toml
[tool.poetry]
name = "flowutils"
version = "0.1.0"
description = ""
authors = ["Denis Stalz-John <denis.stalz-john@codecentric.de>"]
license = "MIT"
readme = "readme.md"

[tool.poetry.dependencies]
python = ">=3.8"
typer = "^0.9.0"
pyyaml = "^6.0.1"
gitpython = "^3.1.33"
rich = "^13.5.2"
pydantic = "^2.3.0"
pydub = "^0.25.1"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
flow = 'flowutils.main:app'
